requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0
pyairtable>=2.0.0
agents>=0.1.0
flask>=2.3.0 
fastapi
uvicorn[standard]
pydantic 